Removing client attribute from priv_queue.
This necessitated also adding a proper qoq_item type for the things
that end up going into the qoq. The qoq_item holds the client now,
so the supplier processor knows not to wake up that client when it
finishes processing the calls.

Having a distinct qoq_item is probably nice anyway, as it means that the
messages do not rely on things like NULL values to delimit the end of a
private queue.